TXNN-15: Implement Cleanup, stage 1 TXNN-22: Implement cleanup post transaction
Motivation
----------
Transactions should not leave metadata around when finished.

Modifications
-------------
* Implemented basic CleanupWorkQueue producer/consumer to cleanup in the background, but decided not to use it yet.  Left in for posterity.
* Implemented end-of-transaction Cleanup via DisposeAsync.

Results
-------
ATR and transaction metadata is cleaned up when DisposeAsync is called on the Transactions object.

Remarks
-------
The data access in the spec is all defined all over the place, and the code reflects this.
Unfortunately, this has led to a good bit of duplication that should really be refactored.
That will be a major refactor, and so I've left it out of this change.

Change-Id: I7a2dc9078a827aa3361ba8075ad22cf3deefc37c
Reviewed-on: http://review.couchbase.org/c/couchbase-transactions-dotnet/+/137958
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>